make timeline layout useable from outside
