+++ src/timebar/index.js
@@ -3,3 +3,4 @@ export { default as TimebarActivity } from './charts/activity'
 export { default as TimebarEvents } from './charts/events'
 export { default as TimebarVesselEvents } from './charts/vessel-events'
 export * from './utils'
+export * from './layouts'
